graphQl/requirements.txt
aniso8601==9.0.1
anyio==4.0.0
ariadne==0.20.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
module-name==0.6.0
packaging==23.2
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.31.1
strawberry-graphql==0.209.8
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
websockets==12.0
Werkzeug==2.3.7
